Make ping use socket marks properly.
ping and ping6 have an option to set the socket mark on the
sockets they use, but unfortunately they do not use that option
on the sockets they use to determine their local address. Thus,
those checks are incorrect, and can result in ping failing to
run if unmarked sockets have no connectivity.

Change-Id: I2b8c30d314d6723a89d26d636f4f3fd0b9b8c3ac